Fix bluetooth service - crashes, pairing, etc.
-add checking permissions before bluetooth actions
-refactoring
-upgrade gradle
-update libs
-change kapt to ksp

+++ app/src/main/java/com/hermanowicz/wirelesswhisper/navigation/features/singleChat/ui/SingleChatScreen.kt
@@ -17,6 +17,7 @@ import androidx.compose.material.TextField
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -26,6 +27,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -34,9 +36,13 @@ import com.hermanowicz.wirelesswhisper.bluetooth.MyBluetoothService
 import com.hermanowicz.wirelesswhisper.components.chatBubble.ChatError
 import com.hermanowicz.wirelesswhisper.components.chatBubble.ChatReceived
 import com.hermanowicz.wirelesswhisper.components.chatBubble.ChatSend
+import com.hermanowicz.wirelesswhisper.components.dialog.DialogPrimary
+import com.hermanowicz.wirelesswhisper.components.permissions.permissionChecker
 import com.hermanowicz.wirelesswhisper.components.topBarScoffold.TopBarScaffold
 import com.hermanowicz.wirelesswhisper.data.model.Message
+import com.hermanowicz.wirelesswhisper.domain.GoToPermissionSettingsUseCase
 import com.hermanowicz.wirelesswhisper.ui.theme.LocalSpacing
+import com.hermanowicz.wirelesswhisper.utils.Permissions
 
 @Composable
 fun SingleChatScreen(
@@ -44,7 +50,42 @@ fun SingleChatScreen(
     viewModel: SingleChatViewModel = hiltViewModel(),
     bluetoothService: Intent
 ) {
-    val uiState by viewModel.uiState.collectAsState()
+    val context = LocalContext.current
+    val uiState by viewModel.state.collectAsState()
+
+    val launcherPermissionsConnectDevice = permissionChecker(
+        onGranted = {
+            sendMessage(
+                context,
+                uiState.currentMessage,
+                bluetoothService,
+                clearTextField = { viewModel.clearCurrentMessage() }
+            )
+        },
+        showDialogPermissionNeeded = { viewModel.showDialogPermissionsSendMessage(true) }
+    )
+
+    if (uiState.showDialogPermissionsSendMessage) {
+        DialogPrimary(
+            onPositiveLabel = stringResource(id = android.R.string.ok),
+            onPositiveRequest = {
+                viewModel.onGoToPermissionSettings(true)
+                viewModel.showDialogPermissionsSendMessage(false)
+            },
+            onDismissRequest = { viewModel.showDialogPermissionsSendMessage(false) }
+        ) {
+            Column(modifier = Modifier.fillMaxWidth()) {
+                Text(text = stringResource(id = R.string.permission_is_needed_to_perform_the_action))
+            }
+        }
+    }
+
+    LaunchedEffect(key1 = uiState.goToPermissionSettings) {
+        if (uiState.goToPermissionSettings) {
+            GoToPermissionSettingsUseCase.invoke(context)
+            viewModel.onGoToPermissionSettings(false)
+        }
+    }
 
     TopBarScaffold(topBarText = uiState.device.name, navigationIcon = {
         IconButton(onClick = navigationIconClick) {
@@ -54,24 +95,24 @@ fun SingleChatScreen(
             )
         }
     }, actions = {
-            Icon(
-                imageVector = Icons.Default.Delete,
-                contentDescription = null,
-                tint = Color.White,
-                modifier = Modifier
-                    .align(Alignment.CenterVertically)
-                    .clickable { viewModel.onClickEditMode(!uiState.deleteMode) }
-            )
-        }) {
+        Icon(
+            imageVector = Icons.Default.Delete,
+            contentDescription = null,
+            tint = Color.White,
+            modifier = Modifier
+                .align(Alignment.CenterVertically)
+                .clickable { viewModel.onClickEditMode(!uiState.deleteMode) }
+        )
+    }) {
         ChatView(
             messageList = uiState.messageList,
-            bluetoothService = bluetoothService,
             currentMessage = uiState.currentMessage,
+            checkPermissions = { launcherPermissionsConnectDevice.launch(Permissions.btPermissions.toTypedArray()) },
             onCurrentMessageChange = {
                 viewModel.onCurrentMessageChange(it)
             },
-            clearTextField = { viewModel.clearCurrentMessage() },
             deleteMode = uiState.deleteMode,
+            setCurrentMessageInState = { viewModel.onCurrentMessageChange(it) },
             onClickDelete = { viewModel.deleteSingleMessage(it.id!!) },
             onCopyMessageToClipboard = { viewModel.copyMessageToClipboard(it) }
         )
@@ -81,11 +122,11 @@ fun SingleChatScreen(
 @Composable
 fun ChatView(
     messageList: List<Message>,
-    bluetoothService: Intent,
     currentMessage: String,
+    checkPermissions: () -> Unit,
     onCurrentMessageChange: (String) -> Unit,
-    clearTextField: () -> Unit,
     deleteMode: Boolean,
+    setCurrentMessageInState: (String) -> Unit,
     onClickDelete: (Message) -> Unit,
     onCopyMessageToClipboard: (String) -> Unit
 ) {
@@ -95,13 +136,18 @@ fun ChatView(
                 .fillMaxWidth()
                 .weight(1f)
         ) {
-            MessagesBox(messageList, deleteMode = deleteMode, onClickDelete = onClickDelete, onCopyMessageToClipboard = onCopyMessageToClipboard)
+            MessagesBox(
+                messageList,
+                deleteMode = deleteMode,
+                onClickDelete = onClickDelete,
+                onCopyMessageToClipboard = onCopyMessageToClipboard
+            )
         }
         MessageBar(
-            bluetoothService,
             currentMessage,
+            checkPermissions,
             onCurrentMessageChange,
-            clearTextField = clearTextField
+            setCurrentMessageInState
         )
     }
 }
@@ -158,13 +204,11 @@ fun MessagesBox(
 
 @Composable
 fun MessageBar(
-    bluetoothService: Intent,
     currentMessage: String,
+    checkPermissions: () -> Unit,
     onCurrentMessageChange: (String) -> Unit,
-    clearTextField: () -> Unit
+    setCurrentMessageInState: (String) -> Unit
 ) {
-    val context = LocalContext.current
-
     Row(
         modifier = Modifier
             .fillMaxWidth()
@@ -179,7 +223,8 @@ fun MessageBar(
             TextField(value = currentMessage, onValueChange = { onCurrentMessageChange(it) })
         }
         IconButton(onClick = {
-            sendMessage(context, currentMessage, bluetoothService, clearTextField = clearTextField)
+            setCurrentMessageInState(currentMessage)
+            checkPermissions()
         }) {
             Icon(painter = painterResource(id = R.drawable.ic_send), contentDescription = null)
         }
